% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistAnimal.R
\docType{data}
\name{DistAnimal}
\alias{DistAnimal}
\title{Henley data (1969)}
\format{
A distance matrix between 12 animals.
}
\usage{
data(DistAnimal)
}
\description{
Mean Interanimal distances from Lists
}
\references{
Henley N. M., (1969). A Psychological Study of the Semantics of Animal Terms. Journal of verbal learning and verbal behavior, 8, 176-184.
}
\keyword{datasets}
