% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_partitioning.R
\name{recursive_partitioning}
\alias{recursive_partitioning}
\title{recursive partitioning}
\usage{
recursive_partitioning(
  d,
  nameobj = NULL,
  crit = "lengthratio",
  crit.ratiodiam = 0.6,
  crit.siz = 3
)
}
\arguments{
\item{d}{: distance matrix for the whole set of objects/items.}

\item{nameobj}{: the label for the objects/items. If NULL (by default) dimnames of d are used.}

\item{crit}{: criterion for splitting, either "lengthratio" (by default) either "proptriplets")}

\item{crit.ratiodiam}{: above the provided threshold the subtree is candidate for splitting. \cr
The criterion considered is the ratio of the diameter of the subset in a subtree \cr
to diameter of the whole sets of objects (relative diameter). \cr
The diameter is defined as the maximum pairwise distance within the set of objects; \cr}

\item{crit.siz}{: size of a subtree (in absolute value), \cr
greater (>) to this threshold (3, by default), the subtree is candidate for splitting.}
}
\value{
\item{partitions}{ Results of the partitions at each step}
}
\description{
recursive partitioning
}
\examples{
data(DistAnimal)
tree=ape::nj(DistAnimal)
ape::plot.phylo(tree,type="unrooted")
recursive_partitioning(d=DistAnimal)

}
\references{
Koenig, L., Cariou, V., Symoneaux, R., Coulon-Leroy, C., Vigneau, E. (2021). Food Quality and Preference, 89, 104137(2021).
}
